feat: add permission sale product

+++ Modules/Sale/Http/Controllers/SaleProductController.php
@@ -5,7 +5,9 @@ namespace Modules\Sale\Http\Controllers;
 use App\Enums\DiscountTarget;
 use App\Enums\DiscountType;
 use App\Enums\TargetType;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Contracts\Support\Renderable;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Controller;
 use Modules\Product\Repositories\ProductRepository;
@@ -14,6 +16,8 @@ use Modules\Sale\Repositories\SaleProductRepository;
 
 class SaleProductController extends Controller
 {
+    use AuthorizesRequests;
+
     /** @var \Modules\Sale\Repositories\SaleProductRepository */
     protected $saleProductRepository;
 
@@ -41,10 +45,18 @@ class SaleProductController extends Controller
      */
     public function index(Request $request, $sale_id)
     {
-        $search = $request->input('search');
-        $sale_products = $this->saleProductRepository->paginateBySaleId($sale_id, $search);
+        try {
+            $this->authorize('sale_product:browse');
 
-        return view('sale::product.index', compact('sale_products', 'sale_id'));
+            $search = $request->input('search');
+            $sale_products = $this->saleProductRepository->paginateBySaleId($sale_id, $search);
+    
+            return view('sale::product.index', compact('sale_products', 'sale_id'));
+        } catch (AuthorizationException $exception) {
+            return redirect()
+            ->route('admin.sale.index')
+            ->with('danger', __('notification.permission.fail', ['action' => 'browse sale product']));
+        }
     }
 
     /**
@@ -53,27 +65,35 @@ class SaleProductController extends Controller
      */
     public function create($sale_id)
     {
-        $form = [
-            'title'     => 'Create',
-            'url'       => route('admin.sale.product.store', $sale_id),
-            'method'    => 'POST',
-        ];
+        try {
+            $this->authorize('sale_product:add');
 
-        $products = $this->productRepository->all();
-        $variants = $this->variantRepository->all();
-        $target_types = TargetType::getObject();
-        $discount_targets = DiscountTarget::getObject();
-        $discount_types = DiscountType::getObject();
-
-        return view('sale::product.create', compact(
-            'form',
-            'products',
-            'variants',
-            'target_types',
-            'discount_targets',
-            'discount_types',
-            'sale_id'
-        ));
+            $form = [
+                'title'     => 'Create',
+                'url'       => route('admin.sale.product.store', $sale_id),
+                'method'    => 'POST',
+            ];
+    
+            $products = $this->productRepository->all();
+            $variants = $this->variantRepository->all();
+            $target_types = TargetType::getObject();
+            $discount_targets = DiscountTarget::getObject();
+            $discount_types = DiscountType::getObject();
+    
+            return view('sale::product.create', compact(
+                'form',
+                'products',
+                'variants',
+                'target_types',
+                'discount_targets',
+                'discount_types',
+                'sale_id'
+            ));
+        } catch (AuthorizationException $exception) {
+            return redirect()
+            ->route('admin.sale.product.index', $sale_id)
+            ->with('danger', __('notification.permission.fail', ['action' => 'add sale product']));
+        }
     }
 
     /**
@@ -83,16 +103,24 @@ class SaleProductController extends Controller
      */
     public function store(Request $request, $sale_id)
     {
-        $request->validate([
-            'target_id'     => 'required',
-            'target_type'   => 'required',
-        ]);
+        try {
+            $this->authorize('sale_product:add');
 
-        $this->saleProductRepository->create($request->all(), ['sale_id' => $sale_id]);
-
-        return redirect()
-        ->route('admin.sale.product.index', $sale_id)
-        ->with('success', __('notification.create.success', ['model' => 'sale product']));
+            $request->validate([
+                'target_id'     => 'required',
+                'target_type'   => 'required',
+            ]);
+    
+            $this->saleProductRepository->create($request->all(), ['sale_id' => $sale_id]);
+    
+            return redirect()
+            ->route('admin.sale.product.index', $sale_id)
+            ->with('success', __('notification.create.success', ['model' => 'sale product']));
+        } catch (AuthorizationException $exception) {
+            return redirect()
+            ->route('admin.sale.product.index', $sale_id)
+            ->with('danger', __('notification.permission.fail', ['action' => 'add sale product']));
+        }
     }
 
     /**
@@ -102,29 +130,37 @@ class SaleProductController extends Controller
      */
     public function edit($sale_id, $id)
     {
-        $form = [
-            'title'     => 'Edit',
-            'url'       => route('admin.sale.product.update', ['sale_id' => $sale_id, 'id' => $id]),
-            'method'    => 'PUT',
-        ];
+        try {
+            $this->authorize('sale_product:edit');
 
-        $products = $this->productRepository->all();
-        $variants = $this->variantRepository->all();
-        $target_types = TargetType::getObject();
-        $discount_targets = DiscountTarget::getObject();
-        $discount_types = DiscountType::getObject();
-        $sale_product = $this->saleProductRepository->find($id);
-
-        return view('sale::product.edit', compact(
-            'form',
-            'products',
-            'variants',
-            'target_types',
-            'discount_targets',
-            'discount_types',
-            'sale_product',
-            'sale_id'
-        ));
+            $form = [
+                'title'     => 'Edit',
+                'url'       => route('admin.sale.product.update', ['sale_id' => $sale_id, 'id' => $id]),
+                'method'    => 'PUT',
+            ];
+    
+            $products = $this->productRepository->all();
+            $variants = $this->variantRepository->all();
+            $target_types = TargetType::getObject();
+            $discount_targets = DiscountTarget::getObject();
+            $discount_types = DiscountType::getObject();
+            $sale_product = $this->saleProductRepository->find($id);
+    
+            return view('sale::product.edit', compact(
+                'form',
+                'products',
+                'variants',
+                'target_types',
+                'discount_targets',
+                'discount_types',
+                'sale_product',
+                'sale_id'
+            ));
+        } catch (AuthorizationException $exception) {
+            return redirect()
+            ->route('admin.sale.product.index', $sale_id)
+            ->with('danger', __('notification.permission.fail', ['action' => 'edit sale product']));
+        }
     }
 
     /**
@@ -135,16 +171,24 @@ class SaleProductController extends Controller
      */
     public function update(Request $request, $sale_id, $id)
     {
-        $request->validate([
-            'target_id'     => 'required',
-            'target_type'   => 'required',
-        ]);
+        try {
+            $this->authorize('sale_product:edit');
 
-        $this->saleProductRepository->update($id, $request->all());
-
-        return redirect()
-        ->route('admin.sale.product.index', $sale_id)
-        ->with('success', __('notification.update.success', ['model' => 'sale product']));
+            $request->validate([
+                'target_id'     => 'required',
+                'target_type'   => 'required',
+            ]);
+    
+            $this->saleProductRepository->update($id, $request->all());
+    
+            return redirect()
+            ->route('admin.sale.product.index', $sale_id)
+            ->with('success', __('notification.update.success', ['model' => 'sale product']));
+        } catch (AuthorizationException $exception) {
+            return redirect()
+            ->route('admin.sale.product.index', $sale_id)
+            ->with('danger', __('notification.permission.fail', ['action' => 'edit sale product']));
+        }
     }
 
     /**
@@ -154,10 +198,18 @@ class SaleProductController extends Controller
      */
     public function destroy($sale_id, $id)
     {
-        $this->saleProductRepository->delete($id);
+        try {
+            $this->authorize('sale_product:delete');
 
-        return redirect()
-        ->route('admin.sale.product.index', $sale_id)
-        ->with('success', __('notification.delete.success', ['model' => 'sale product']));
+            $this->saleProductRepository->delete($id);
+    
+            return redirect()
+            ->route('admin.sale.product.index', $sale_id)
+            ->with('success', __('notification.delete.success', ['model' => 'sale product']));
+        } catch (AuthorizationException $exception) {
+            return redirect()
+            ->route('admin.sale.product.index', $sale_id)
+            ->with('danger', __('notification.permission.fail', ['action' => 'delete sale product']));
+        }
     }
 }
